WIP - Fix failing tests - continued
